Use int for time.

+++ tests/test_simulator.py
@@ -12,7 +12,7 @@ from workload import Job, Resource, Resources, TaskGraph, TaskState
 class MockScheduler(BaseScheduler):
     """ A MockScheduler that enables the testing of the Simulator. """
 
-    def __init__(self, runtime: float, placement):
+    def __init__(self, runtime: int, placement):
         self._runtime = runtime
         self._task_placement = placement
         self.preemptive = False
@@ -50,24 +50,23 @@ def __create_default_worker_pool(
 
 def test_event_construction_success():
     """ Test that an Event can be successfully constructed. """
-    event = Event(event_type=EventType.SCHEDULER_START, time=1.0)
+    event = Event(event_type=EventType.SCHEDULER_START, time=1)
     assert event.event_type == EventType.SCHEDULER_START, "Invalid Event type."
-    assert event.time == 1.0, "Invalid event time."
+    assert event.time == 1, "Invalid event time."
 
 
 def test_event_construction_failure():
     """ Test that an Event without a task fails to be constructed. """
     with pytest.raises(ValueError):
-        Event(event_type=EventType.TASK_RELEASE, time=1.0)
+        Event(event_type=EventType.TASK_RELEASE, time=1)
 
 
 def test_event_queue_one():
     """ Test that an EventQueue properly orders the given events. """
     event_queue = EventQueue()
-    event_queue.add_event(Event(event_type=EventType.SCHEDULER_START,
-                                time=1.0))
+    event_queue.add_event(Event(event_type=EventType.SCHEDULER_START, time=1))
     event_queue.add_event(
-        Event(event_type=EventType.SCHEDULER_FINISHED, time=2.0))
+        Event(event_type=EventType.SCHEDULER_FINISHED, time=2))
 
     assert len(event_queue) == 2, "Incorrect number of events in the queue."
     assert event_queue.next().event_type == EventType.SCHEDULER_START,\
@@ -83,12 +82,11 @@ def test_event_queue_two():
     event_queue = EventQueue()
     event_queue.add_event(
         Event(event_type=EventType.TASK_RELEASE,
-              time=3.0,
+              time=3,
               task=create_default_task()))
     event_queue.add_event(
-        Event(event_type=EventType.SCHEDULER_FINISHED, time=2.0))
-    event_queue.add_event(Event(event_type=EventType.SCHEDULER_START,
-                                time=1.0))
+        Event(event_type=EventType.SCHEDULER_FINISHED, time=2))
+    event_queue.add_event(Event(event_type=EventType.SCHEDULER_START, time=1))
 
     assert len(event_queue) == 3, "Incorrect number of events in the queue."
     assert event_queue.next().event_type == EventType.SCHEDULER_START,\
@@ -107,7 +105,7 @@ def test_simulator_construction():
     """ Test that a simulator can be correctly constructed. """
     worker_pool = __create_default_worker_pool()
     simulator = Simulator(worker_pools=[worker_pool],
-                          scheduler=MockScheduler(runtime=1.0, placement=None),
+                          scheduler=MockScheduler(runtime=1, placement=None),
                           job_graph=None)
     assert len(simulator._worker_pools) == 1, "Incorrect number of WorkerPool"
     assert len(simulator._event_queue) == 2,\
@@ -118,45 +116,45 @@ def test_failed_construction_of_scheduler_start_event():
     """ Test if a non SCHEDULER_FINISHED event raises an error or not. """
     worker_pool = __create_default_worker_pool()
     simulator = Simulator(worker_pools=[worker_pool],
-                          scheduler=MockScheduler(runtime=1.0, placement=None),
+                          scheduler=MockScheduler(runtime=1, placement=None),
                           job_graph=None)
     with pytest.raises(ValueError):
         simulator._Simulator__get_next_scheduler_event(
-            event=Event(event_type=EventType.SCHEDULER_START, time=3.0),
+            event=Event(event_type=EventType.SCHEDULER_START, time=3),
             task_graph=TaskGraph(),
-            scheduler_frequency=-1.0,
-            last_scheduler_start_time=2.0)
+            scheduler_frequency=-1,
+            last_scheduler_start_time=2)
 
 
 def test_construction_of_scheduler_start_event():
     """ Test the correct construction of a SCHEDULER_START event. """
     worker_pool = __create_default_worker_pool()
     simulator = Simulator(worker_pools=[worker_pool],
-                          scheduler=MockScheduler(runtime=1.0, placement=None),
+                          scheduler=MockScheduler(runtime=1, placement=None),
                           job_graph=None)
 
     simulator_start_event = simulator._Simulator__get_next_scheduler_event(
-        event=Event(event_type=EventType.SCHEDULER_FINISHED, time=3.0),
+        event=Event(event_type=EventType.SCHEDULER_FINISHED, time=3),
         task_graph=TaskGraph(),
-        scheduler_frequency=-1.0,
-        last_scheduler_start_time=1.0)
-    assert simulator_start_event.time == 4.0,\
+        scheduler_frequency=-1,
+        last_scheduler_start_time=1)
+    assert simulator_start_event.time == 4,\
         "Incorrect start time for Scheduler."
 
     simulator_start_event = simulator._Simulator__get_next_scheduler_event(
-        event=Event(event_type=EventType.SCHEDULER_FINISHED, time=3.0),
+        event=Event(event_type=EventType.SCHEDULER_FINISHED, time=3),
         task_graph=TaskGraph(),
-        scheduler_frequency=5.0,
-        last_scheduler_start_time=1.0)
-    assert simulator_start_event.time == 6.0,\
+        scheduler_frequency=5,
+        last_scheduler_start_time=1)
+    assert simulator_start_event.time == 6,\
         "Incorrect start time for Scheduler."
 
     simulator_start_event = simulator._Simulator__get_next_scheduler_event(
-        event=Event(event_type=EventType.SCHEDULER_FINISHED, time=7.0),
+        event=Event(event_type=EventType.SCHEDULER_FINISHED, time=7),
         task_graph=TaskGraph(),
-        scheduler_frequency=5.0,
-        last_scheduler_start_time=1.0)
-    assert simulator_start_event.time == 8.0,\
+        scheduler_frequency=5,
+        last_scheduler_start_time=1)
+    assert simulator_start_event.time == 8,\
         "Incorrect start time for Scheduler."
 
 
@@ -164,7 +162,7 @@ def test_simulator_loop_finish_event():
     """ Test that a correct SIMULATOR_END event is generated. """
     worker_pool = __create_default_worker_pool()
     simulator = Simulator(worker_pools=[worker_pool],
-                          scheduler=MockScheduler(runtime=1.0, placement=None),
+                          scheduler=MockScheduler(runtime=1, placement=None),
                           job_graph=None)
     assert simulator._event_queue.next().event_type ==\
         EventType.SIMULATOR_START, "Incorrect event received."
@@ -174,10 +172,10 @@ def test_simulator_loop_finish_event():
         "Incorrect length for event queue."
 
     simulator_start_event = simulator._Simulator__get_next_scheduler_event(
-        event=Event(event_type=EventType.SCHEDULER_FINISHED, time=3.0),
+        event=Event(event_type=EventType.SCHEDULER_FINISHED, time=3),
         task_graph=TaskGraph(),
-        scheduler_frequency=-1.0,
-        last_scheduler_start_time=1.0)
+        scheduler_frequency=-1,
+        last_scheduler_start_time=1)
     assert simulator_start_event.event_type == EventType.SIMULATOR_END,\
         "Incorrect event received from next_scheduler_event"
 
@@ -186,27 +184,27 @@ def test_scheduler_invocation_by_simulator():
     """ Test that the simulator correctly invokes the scheduler. """
     worker_pool = __create_default_worker_pool()
     simulator = Simulator(worker_pools=[worker_pool],
-                          scheduler=MockScheduler(runtime=5.0, placement=None),
+                          scheduler=MockScheduler(runtime=5, placement=None),
                           job_graph=None)
     scheduler_finished_event = simulator._Simulator__run_scheduler(
-        event=Event(event_type=EventType.SCHEDULER_START, time=1.0),
+        event=Event(event_type=EventType.SCHEDULER_START, time=1),
         task_graph=TaskGraph(),
     )
-    assert scheduler_finished_event.time == 6.0, "Incorrect finish time."
+    assert scheduler_finished_event.time == 6, "Incorrect finish time."
 
 
 def test_simulator_step():
     """ Test that the Simulator's step() method correctly inserts events. """
     worker_pool = __create_default_worker_pool()
     simulator = Simulator(worker_pools=[worker_pool],
-                          scheduler=MockScheduler(runtime=5.0, placement=None),
+                          scheduler=MockScheduler(runtime=5, placement=None),
                           job_graph=None)
 
     # Create, release and place a Task.
-    task = create_default_task(runtime=3.0)
-    task.release(1.0)
+    task = create_default_task(runtime=3)
+    task.release(1)
     worker_pool.place_task(task)
-    task.start(2.0)
+    task.start(2)
 
     assert simulator._simulator_time == 0, "Incorrect starting simulator time."
     assert len(simulator._event_queue) == 2,\
@@ -250,18 +248,18 @@ def test_simulator_handle_event():
     """ Test the Simulator's handle_event method with different events. """
     worker_pool = __create_default_worker_pool()
     simulator = Simulator(worker_pools=[worker_pool],
-                          scheduler=MockScheduler(runtime=5.0, placement=None),
+                          scheduler=MockScheduler(runtime=5, placement=None),
                           job_graph=None)
 
     # Test the SIMULATOR_START event.
     return_value = simulator._Simulator__handle_event(event=Event(
-        event_type=EventType.SIMULATOR_START, time=1.0),
+        event_type=EventType.SIMULATOR_START, time=1),
                                                       task_graph=TaskGraph())
     assert not return_value, "Incorrect return value for event type."
 
     # Test the SIMULATOR_END event.
     return_value = simulator._Simulator__handle_event(event=Event(
-        event_type=EventType.SIMULATOR_END, time=1.0),
+        event_type=EventType.SIMULATOR_END, time=1),
                                                       task_graph=TaskGraph())
     assert return_value, "Incorrect return value for event type."
 
@@ -270,14 +268,14 @@ def test_simulator_handle_event():
     planning_task = create_default_task(job=Job(name="Planning"))
     task_graph = TaskGraph()
     task_graph.add_task(perception_task, [planning_task])
-    perception_task.release(2.0)
-    perception_task.start(3.0)
+    perception_task.release(2)
+    perception_task.start(3)
     perception_task.update_remaining_time(0)
-    perception_task.finish(4.0)
+    perception_task.finish(4)
 
     assert len(simulator._event_queue) == 2, "Incorrect length of EventQueue."
     return_value = simulator._Simulator__handle_event(event=Event(
-        event_type=EventType.TASK_FINISHED, time=4.0, task=perception_task),
+        event_type=EventType.TASK_FINISHED, time=4, task=perception_task),
                                                       task_graph=task_graph)
     assert not return_value, "Incorrect return value for event type."
     assert len(simulator._event_queue) == 3, "Incorrect length of EventQueue."
@@ -285,7 +283,7 @@ def test_simulator_handle_event():
     # Test the SCHEDULER_START event.
     return_value = simulator._Simulator__handle_event(event=Event(
         event_type=EventType.SCHEDULER_START,
-        time=5.0,
+        time=5,
     ),
                                                       task_graph=TaskGraph())
     assert not return_value, "Incorrect return value for event type."
@@ -295,7 +293,7 @@ def test_simulator_handle_event():
     simulator._last_task_placement = [(planning_task, worker_pool.id)]
     return_value = simulator._Simulator__handle_event(event=Event(
         event_type=EventType.SCHEDULER_FINISHED,
-        time=6.0,
+        time=6,
     ),
                                                       task_graph=TaskGraph())
     assert len(simulator._event_queue) == 5, "Incorrect length of EventQueue."
